Fix bug around changing the organisation for a request
If you used the Edit form to change the organisation associated with a request,
it would change the organisation_slug stored in the database, but it would
appear not to have worked, because everywhere else the old title would be
displayed.

This is because the model wouldn't update its denormalised copy of the
organisation title, unless the organisation title was blank.

This wasn't an issue when the organisation was only set once - at create-time.

I had to tweak the factory so that it could continue to work.

A better fix would be to normalise the data, but that's a much bigger change.