Add migration to remove redirect
This redirect was put in place as a temporary measure while a new Smart Answer
was being developed for HMRC.

The new Smart Answer (part-year-profit-tax-credits)[1] was deployed on Tue 29
Sep. Before the deployment we had to change the owner of this path in
url-arbiter from short-url-manager to smartanswers[2].

@jamiecobbett mentioned that we should also remove the redirect from the
short-url-manager database now that it's no longer required.

I've followed the pattern in previous migrations of finding the
`ShortUrlRequest` and `Redirect` records we're interested in.

[1]: https://github.com/alphagov/smart-answers/pull/1942
[2]: https://github.com/alphagov/url-arbiter/pull/25