Refactor all short url request operations to commands
The intention of this commit is to move all the business logic from
hooks to a command object.

This allows a developer to see, in a single place, what happens when
each operation is performed, rather than having the behaviour of the
application implemented as 'side effects' that have to be traced in
multiple places - the controller, the model, and its associated
redirect.

Subsequent refactors are likely to include ensuring the publishing API
behaviour moving into these classes too as operations that occur during
the lifecycle, but for now I'm leaving them in place as I'm not touching
that class.

I've also changed the rails helper to stub all requests to the
publishing api in all tests; this is mainly because so many tests
started to require it (an 'accepted' request must have a redirect which
then forces a call to the publishing api) that it was adding noise to
the spec suites.

This refactor will remove the need for this once the side effect
behaviour is removed.