Fail validation if an identical redirect exists
We don't want to create extra work for reviewers -- prevent raising a
request if a current Redirect already exists with the same from_path and
to_path.

I spotted a bug here in that if validation failed but confirmation was
required the validation wouldn't be checked until the confirmation was
sent, at which point the user is redirected to the form, and the
confirmation is lost, so they have to confirm again. Fix that by
massaging the order (and cover it with a spec).