Mark any existing requests as superseded when accepting a request
In 52198695ffb39695dd8d98adc246b91721ea8bdc we stopped the system from
destroying an existing request when we accept a new one for the same
from_path.  This leaves the existing request in the DB in the 'accepted'
state, but without an attachment to a Redirect instance.  When we view
live requests we load all 'accepted' requests and so we will show the
user requests that have been replaced.

Instead of destroying the existing request we mark them as superseded.
This means they won't appear on the the live requests page, but also that
they'll appear in the "Other requests" list as an audit trail.

Note that in the Accept command we have to use the target of the relation
instead of the relation directly to update the state.  This is because all
relation objects in mongoid are proxies to the DB and not the actual
object. If the data for the relation changes in the DB the proxy will
update when you next reference it to show that change.  In this case we
don't want that, so we have to use the target which is the real object
rather than the proxy.

Note that this is true in mongoid 4 at least (see last paragraph before
the navigation on
https://mongoid.github.io/old/en/mongoid/docs/relations.html).  We should
revisit this when we upgrade to mongoid 5 or beyond.