Stop using ci_reporter
This is a gem only used in the test environment but we require it all
the time.  When deploying the app we don't install this gem so it's
not available and this breaks running rake tasks on deploy:

      * executing "cd /data/apps/short-url-manager/releases/20170116103020 && govuk_setenv short-url-manager bundle exec rake RAILS_ENV=production RAILS_GROUPS=assets assets:precompile"
        servers: ["backend-1.backend.integration.publishing.service.gov.uk", "backend-2.backend.integration.publishing.service.gov.uk"]
        [backend-2.backend.integration.publishing.service.gov.uk] executing command
        [backend-1.backend.integration.publishing.service.gov.uk] executing command
    *** [err :: backend-2.backend.integration.publishing.service.gov.uk] rake aborted!
    *** [err :: backend-2.backend.integration.publishing.service.gov.uk] LoadError: cannot load such file -- ci/reporter/rake/rspec

We could conditionally require the ci/reporter/rake/rspec file, or, as
it doesn't really seem to do anything useful, we can just stop using it
entirely.