Use correct host for links to short url requests in notification emails
These urls are hosted on the short-url-manager domain, not the public www
domain.  We could solve this by setting default_url_options on the mailer
but as we generat two types of urls (public and admin) it's not obvious
which host should be the default.  It's probably better to set neither
and force us to be specific when generating all urls.