Require signin permission for all visitors
Previously we would allow any signon user to log in to the app even if
they didn't have 'signin' permission.  This could lead to unexpected
results where we uncheck the "access" permission in the signon app, but
the user is still able to login and this is not what we expect.

We can't write a test for this because in development and test we use the
`mock_gds_sso` strategy which adds the `signin` permission to every user
that logs in (see:
https://github.com/alphagov/gds-sso/blob/4ca7fe45c823f6a1684b17933f83f1d4aa674c6d/lib/gds-sso/warden_config.rb#L86-L89)
- meaning that we can't try to login as someone without `signin`
permission.