Add rake tasks to set default values
Redirect now has 2 new fields from previous commits, but these were not
passing values to the publishing-api if the fields were not in the
mongo document.

We tried db migrations but hit an error because the version we have for
mongoid_rails_migrations does not support Mongo 5 upwards. The project
has a fix in their master branch, but have not yet released it.

The migrations works if we point to the master branch - but we have
decided against doing this moving forward.

Instead we add one-off rake tasks to set default values for these fields.