Require a specific permission for using new functionality
Require a specific permission to make and accept redirect requests
that use non-default options for the route type and segments mode.

This is to allow deploying these changes, while retaining temporary
control over who is able to use these changes.

This is to support the migration of router-data content to the
Publishing API, as the model changes being un-reverted support the
full range of options that are needed for this migration.