Add a csv file with redirects from router-data
This csv file was generated by selecting data from router-data where
there is nothing in the Publishing API preventing the redirects from
being created.

The following query was used [1], the full steps to run this are explained
in the migration report (migration-plan.org in the router-data
repository).

This can be read by the redirects:import rake task, which will create
redirects in the Short URL Manager and Publishing API for each line in
the file.

1:
\copy (WITH live_editions AS (SELECT base_path, publishing_app FROM editions WHERE content_store = 'live')  SELECT router_data.* FROM router_data  WHERE NOT EXISTS (SELECT 1 FROM live_editions WHERE base_path = source) AND NOT EXISTS (SELECT 1 FROM path_reservations WHERE base_path = source)) to '/tmp/not-in-publishing-api.csv' with CSV header