Add not-in-publishing-api-fix-segments-mode.csv
This is a subset of the redirects in
router-data-import/not-in-publishing-api.csv, where the route type is
prefix, and the segments mode is unspecified. These should have a
segments_mode of preserve, but due to the import rake task not
computing the default, are incorrectly set to exact.

This file is being added so that these redirects can be re-imported
once the import rake task is fixed, to fix the data in the Publishing API.