Fix Brakeman SQL Injection warning
We currently get a warning from Brakeman for possible SQL injection from
the organisations importer's redis-lock that is blocking other PRs from
being merged.

This was added in https://github.com/alphagov/short-url-manager/pull/8
but appears to now be unnecessary as the Rails environment is always
`production` in non-test and non-CI environments anyway and each
environment has it's own Redis instance.