Unreserve path when redirect deleted
When a redirect is created using Short URL Manager, the ownership of
that base path changes from the original publishing application to
"short-url-manager" in Publishing API.

Currently, when the redirect is deleted, the unpublish type is set to
"gone" rather than "redirect", but the base path is still owned by
Short URL Manager, meaning that the original page cannot be republished,
nor can the base path be used for a new page unless a developer brings
up a rails console and manually calls the "unreserve_path" method.

Calling "unreserve_path" when the redirect is deleted removes this need
for manual intervention.