Add optimalworkshop.com to allowlist of external domains
The explore team is adding a recruitment banner to subsets of pages
on govuk. The banner contains a link to the online survey at the
above domain. To save us having to redeploy the frontend applications
every time the survey URL is updated we would like to set up a redirect
in short-url manager. Once we have completed this phase of testing
optimalworkshop.com will be removed from the allowlist.