Fix compilation error
When adding the govuk publishing components gem into the dependencies
for this app so the google tag manager snippet could be used we started
encountering this error:

```SassC::SyntaxError: Internal Error: Incompatible units: 'px' and 'rem'.```

This was caused by the compression we were doing in the test env.

We have now removed the compression.