Fix links to absolute URLs
govuk_url_for is used in three places in this codebase, all in view code
like:

      <%= link_to short_url_request.to_path, govuk_url_for(short_url_request.to_path) %>

`to_path` can be an absolute URL though, and in this case the
application will render a link like:

     <a href="http://gov.ukhttp://example.com">http://example.com</a>

which is real clown shoes stuff.

Addressable's URI.join method basically does what we want - if we're
given a path, it will join that on to the website root, if we're given
an absolute URL it will give us that.