WIP: catch "route already exists in Publishing API" error
We had a situation earlier whereby Short URL Manager tried and
failed to apply a redirect to an existing route, despite the
'override existing' field set to 'no'. The response was a 422 and
no error message, which meant the user had to get in touch with
support.

We should investigate and fix the 'override existing' check,
as well as do a better job at surfacing errors.

I tried and failed to recreate the scenario here in several different
test files, but it may be useful as a starting point for whoever
picks up the card.

Trello: https://trello.com/c/PVcAt5Ll/3178-surface-path-reserved-errors-in-ui