Remove rake task that relies on CSV input
Rake tasks that rely on CSV input are a throwback to the old
Puppet infrastructure of GOV.UK, where each app was running on
dedicated VMs, meaning there was always a fairly stable/reliable
file system to which we could upload said CSV and then run the
rake task against it via Jenkins. Kubernetes pods, in contrast,
are ephemeral in nature, so any uploaded CSVs could be lost at
any time. We should therefore migrate CSV-based rake tasks into
the UI of the respective publishing app, or remove it if it is no
longer needed.

This particular rake task was added in July 2017 in #86, with the
comment "This is a one off task for importing data from
router-data". It is therefore safe to assume it is no longer
needed and can be deleted.

Saying that, it was used - once - in May 2021, in #601. If we ever
need to import redirects in bulk again, we can add this code back
in if needed, but what's really needed longer term is a bulk
import feature in the UI, which we should consider the next time
the need arises.

Trello: https://trello.com/c/yi1ExGvz/3179-audit-rake-tasks-that-rely-on-scp-push