Create task for restoring missing redirect content IDs from Publishing API
780 redirects seem to be missing content IDs. Occasionally departments will request that we reuse one of the redirected paths for a new document on GOV.UK, so we need to unpublish the redirects.

However, this isn't possible if the content ID is missing because Publishing API's unpublishing API endpoint requires the client to send a content ID.

Fortunately, Publishing API also provides an endpoint for looking up the content ID if one has a base path, so we can find the content ID for each redirect and save it in our database. This will enable us to unpublish these redirects in future, should it be necessary.