Discard drafts when unpublishing redirects
Short URL manager doesn't support "drafts", but there seems to be a number of redirects in Publishing API that do have drafts. Publishing API will refuse to unpublish redirects in Publishing API that have a draft, unless the `discard_drafts` option is passed in the API request so that the draft can be discarded.